[FLINK-21377][coordination][tests] Move JobMaster QS tests to separate file
